requirements.txt
Flask==2.3.3
pytesseract==0.3.10
Pillow==10.0.0
llama-cpp-python>=0.3.0
pytest>=8.1.0
playwright>=1.44.0
pytest-playwright>=0.5.0
pytest-html>=3.2.0
